Refactor CheckHook to HookList for more accurate reusability
