+++ cli/commands/check/list.go
@@ -95,7 +95,7 @@ func printToTable(results interface{}, writer io.Writer) {
 			Title: "Hooks",
 			CellTransformer: func(data interface{}) string {
 				check, _ := data.(types.CheckConfig)
-				return globals.FormatCheckHooks(check.CheckHooks)
+				return globals.FormatHookLists(check.CheckHooks)
 			},
 		},
 		{
