+++ types/check.proto
@@ -70,18 +70,10 @@ message CheckConfig {
   string source = 13;
 
 	// CheckHooks is the list of check hooks for the check
-	repeated CheckHook check_hooks = 14 [(gogoproto.jsontag) = "check_hooks", (gogoproto.nullable) = false];
+	repeated HookList check_hooks = 14 [(gogoproto.jsontag) = "check_hooks", (gogoproto.nullable) = false];
 
   // STDIN indicates if the check command accepts JSON via stdin from the agent
   bool stdin = 15;
-}
-
-message CheckHook {
-	// Hooks is the list of hooks for the check hook
-	repeated string hooks = 1 [(gogoproto.jsontag) = "hooks"];
-
-	// Type indicates the type or response code for the check hook
-	string type = 2;
 }
 
 // A Check is a check specification and optionally the results of the check's
