+++ types/hook.proto
@@ -60,3 +60,11 @@ message Hook {
 	// Status is the exit status code produced by the hook
 	int32 status = 6;
 }
+
+message HookList {
+	// Hooks is the list of hooks for the check hook
+	repeated string hooks = 1 [(gogoproto.jsontag) = "hooks"];
+
+	// Type indicates the type or response code for the check hook
+	string type = 2;
+}
